[Рейтинг] в реквест может быть дробное значение

+++ src/main/java/club/ttg/dnd5/domain/common/model/Rating.java
@@ -20,5 +20,5 @@ public class Rating {
     private String url;
     @CreatedBy
     private String username;
-    private byte value;
+    private double value;
 }
